# openapi.yaml
openapi: 3.0.3
info:
  title: BarberPOS Backend API
  version: 0.1.0
servers:
  - url: https://api.barberpos.cloud
paths:
...
  /products:
    get:
      summary: List products/services
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
    post:
      summary: Upsert product (manager/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
  /products/{id}:
    delete:
      summary: Soft delete product (manager/admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          ok: { type: boolean }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: integer }
        status: { type: string }
    ApiResponse:
      type: object
      required: [status, data]
      properties:
        status: { type: string, enum: [success, error] }
        message: { type: string }
        data: {}
        error:
          $ref: '#/components/schemas/ApiError'
...
    Product:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
        price: { type: integer }
        image: { type: string }
        trackStock: { type: boolean }
        stock: { type: integer }
        minStock: { type: integer }
...